Don't try to remove '.' and '..'.
cleanup() is supposed to remove files from a temporary directory and
then remove the temporary directory.

+++ sys/tools/makesyscalls.lua
@@ -110,7 +110,9 @@ local function cleanup()
 	if cleantmp then
 		if lfs.dir(tmpspace) then
 			for fname in lfs.dir(tmpspace) do
-				os.remove(tmpspace .. "/" .. fname)
+				if fname ~= "." and fname ~= ".." then
+					os.remove(tmpspace .. "/" .. fname)
+				end
 			end
 		end
